layout: replace unnecessary <div class="post"> with <article>
There's no styling for it. I'm unsure why I have it. Instead, make
<article> the containing element.

This feels semantically more correct: a title is part of the article, so
is its publishing date. See
https://stackoverflow.com/questions/59564236/semantic-markup-h1-in-page-header-or-article-header.

Also wrap the tags block in an <aside>, which indicates (semantically)
that it's not part of the main article. I haven't decided yet whether I
want this to be part of <article> at some point too.